Refactored player name editing by moving `NameEditDialog` to a dedicated component and relocating its trigger from `MainScreen` to `ProfileScreen`. Updated the `MainScreen` UI with a slight adjustment to the title font size and removed unused name-related logic and state.
